feat(bundle,pdf,search): handle nullable selected funds, fix PDF crash/download, in-place MF search
Bundle — nullable selected funds:
- PortfolioSlotDomain.selectedFund is now nullable (FundDomain?); mapper uses
  getOrNull(index) instead of falling back to the first fund.
- BundleReviewScreen: SlotReviewItem renders an "empty slot" card ("No fund
  selected for this category") when selectedFund is null; Proceed button enables
  only when every slot across all categories has a fund; info box shows the real
  selected/total count. (Includes a debug LaunchedEffect that logs bundle rules.)
- SelectFundScreen: slot chip shows "--" when no fund; pre-selected section shows
  "No fund has selected previously"; Save button enables only when the active
  slot's fund is non-null; null-safe filtering of similar funds / other slots.
- BundleDetailsViewModel: blocks add-to-cart with a snackbar if any slot is
  unselected; BundleSelection.allocation_percentage is now Double.
- deriveTransactionRules / calculateBundleMinimum: null-safe via mapNotNull +
  reduceOrNull over selected funds only.
- BundleDetailsScreen / ExploreCategoryFundScreen: null-safe fund name display
  and minor label/color tweaks.

PDF (Android):
- PdfViewer.android: guard startActivity against ActivityNotFoundException with a
  browser fallback and a toast, fixing the Android 13 open-PDF crash.
- PdfDownloader.downloadUrlPdf: use DownloadManager (public Downloads dir, system
  completion notification that opens the PDF), polling for progress/success.

Mutual fund search:
- Pressing search now reloads results in place (discard funds + reload page 1)
  keeping search text and filters, instead of navigating to a new screen.
- Fixed loadNext() to paginate with the current search text.

Other:
- Portfolio: FolioFundScreen route carries actualFolio; "Current Value" label;
  card icon shadows. Goals: toggle selection keyed by folio. FD: rounded rate.
  Cart cutoff popup defaults consent checked + fully expanded. HttpClient: uses a
  hardcoded bearer token for testing.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/data/remote/model/bundlecart/BundleCartRequest.kt
@@ -26,5 +26,5 @@ data class AddBundleSipRequest(
 @Serializable
 data class BundleSelection(
     val mf_product_id: String,
-    val allocation_percentage: Int
+    val allocation_percentage: Double
 )
